+++ dart/test/http_client/tracing_client_test.dart
@@ -160,13 +160,38 @@ void main() {
       final response = await sut.get(requestUri);
 
       final baggageHeader = propagationContext.toBaggageHeader();
-      final sentryTraceHeader = propagationContext.toSentryTrace();
 
       expect(propagationContext.toBaggageHeader(), isNotNull);
       expect(
           response.request!.headers[baggageHeader!.name], baggageHeader.value);
-      expect(response.request!.headers[sentryTraceHeader.name],
-          sentryTraceHeader.value);
+
+      final traceHeader = SentryTraceHeader.fromTraceHeader(
+        response.request!.headers['sentry-trace'] as String,
+      );
+      expect(traceHeader.traceId, propagationContext.traceId);
+      // can't check span id as it is always generated new
+    });
+
+    test(
+        'tracing header from propagation context should generate new span ids for new events',
+        () async {
+      fixture._hub.options.tracesSampleRate = null;
+      fixture._hub.options.tracesSampler = null;
+      final sut = fixture.getSut(
+        client: fixture.getClient(statusCode: 200, reason: 'OK'),
+      );
+      final propagationContext = fixture._hub.scope.propagationContext;
+      propagationContext.baggage = SentryBaggage({'foo': 'bar'});
+
+      final response1 = await sut.get(requestUri);
+      final response2 = await sut.get(requestUri);
+
+      final traceHeader1 = SentryTraceHeader.fromTraceHeader(
+          response1.request!.headers['sentry-trace']!);
+      final traceHeader2 = SentryTraceHeader.fromTraceHeader(
+          response2.request!.headers['sentry-trace']!);
+
+      expect(traceHeader1.spanId, isNot(traceHeader2.spanId));
     });
 
     test(